+++ src/components/Header/components/Mobile.tsx
@@ -1,18 +1,21 @@
 import React from "react";
-import { chakra, Flex, HStack, useMultiStyleConfig } from "@chakra-ui/react";
+import { chakra, Flex, HStack } from "@chakra-ui/react";
 
 import Icon from "../../Icons/Icon";
+import LoginButton from "./LoginButton";
+import { LoginProps } from "./UpperNav";
+import MobileNavButton from "./MobileNavButton";
+import SearchButton from "./SearchButton";
 
-const Mobile = chakra(() => {
-  const styles = useMultiStyleConfig("HeaderMobile", {});
-  return (
+const Mobile = chakra(
+  ({ isLoginOpen, patronName, setIsLoginOpen }: LoginProps) => (
     <Flex>
-      <HStack __css={styles.icons}>
-        <Icon
-          id="log-in-icon"
-          name="utilityAccountUnfilled"
-          size="medium"
-          title="Log in to your account"
+      <HStack>
+        <LoginButton
+          isMobile
+          isLoginOpen={isLoginOpen}
+          patronName={patronName}
+          setIsLoginOpen={setIsLoginOpen}
         />
         <Icon
           id="locator-icon"
@@ -20,16 +23,11 @@ const Mobile = chakra(() => {
           size="medium"
           title="NYPL Locator"
         />
-        <Icon id="search-icon" name="search" size="medium" title="Search" />
-        <Icon
-          id="hamburger-icon"
-          name="utilityHamburger"
-          size="medium"
-          title="Menu"
-        />
+        <SearchButton isMobile />
+        <MobileNavButton />
       </HStack>
     </Flex>
-  );
-});
+  )
+);
 
 export default Mobile;