+++ src/components/Header/components/UpperNav.tsx
@@ -3,30 +3,33 @@ import { chakra, useMultiStyleConfig } from "@chakra-ui/react";
 
 import Button from "../../Button/Button";
 import Icon from "../../Icons/Icon";
-import Login from "./Login";
 import Link from "../../Link/Link";
 import List from "../../List/List";
+import LoginButton from "./LoginButton";
+
 export interface LoginProps {
-  loginOpen: boolean;
+  catalogRef?: React.RefObject<HTMLAnchorElement>;
+  greetingRef?: React.RefObject<HTMLDivElement>;
+  isMobile?: boolean;
+  isLoginOpen?: boolean;
   patronName: string;
-  setLoginOpen: (boolean) => void;
+  setIsLoginOpen?: (boolean) => void;
 }
 
 const UpperNav = chakra(
-  ({ loginOpen, patronName, setLoginOpen }: LoginProps) => {
+  ({ isLoginOpen, patronName, setIsLoginOpen }: LoginProps) => {
     const styles = useMultiStyleConfig("HeaderUpperNav", {});
     return (
       <nav aria-label="Header top links">
         <List
-          __css={styles.upperNav}
           id="header-nav-upper"
           inline
           listItems={[
-            <Login
+            <LoginButton
               key="login"
-              loginOpen={loginOpen}
+              isLoginOpen={isLoginOpen}
               patronName={patronName}
-              setLoginOpen={setLoginOpen}
+              setIsLoginOpen={setIsLoginOpen}
             />,
             <Link href="#" key="locationsLink">
               Locations
@@ -34,12 +37,7 @@ const UpperNav = chakra(
             <Link href="#" key="libraryCardLink">
               Get A Library Card
             </Link>,
-            <Link
-              className="withIcon"
-              id="emailUpdatesButton"
-              key="emailUpdatesButton"
-              href="#"
-            >
+            <Link id="emailUpdatesButton" key="emailUpdatesButton" href="#">
               <>
                 Get Email Updates
                 <Icon name="arrow" align="right" size="small" />
@@ -54,6 +52,7 @@ const UpperNav = chakra(
           ]}
           noStyling
           type="ul"
+          __css={styles.upperNav}
         />
       </nav>
     );